dependabot: Remove /tools configuration
The tools directory is being removed in the very near future as things
move over to twoliter. Dependencies will be updated there, so any
updates now are questionable since they add to the churn as we try to
live migrate over to the new location.

This removes the tools-specific configuration from the dependabot config
to try to reduce the noise.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>