skip reinstillation of twoliter from source
When we are testing changes to Twoliter, we need to install from git. In
that case the sha is not going to match the version output from Twoliter
--version. This means that we have to rebuild and install Twoliter each
time we call cargo make. It takes 10 minutes to compile Twoliter.

Fix this such that TWOLITER_ALLOW_BINARY_INSTALL=false skips checking
the version when any `twoliter` is installed.