tools/diff-kernel-config: Write summary to file
Write out the overview of config changes to a file in addition to
showing it on the console. This can come in handy when using this script
as part of other automation to recall that information in a final bigger
report. An example application could be automatic composition of PR
cover letters.

Signed-off-by: Leonard Foerster <foersleo@amazon.com>