kernel-6.1: cherry-pick fix for creating kprobes using unqualified names
Commit b022f0c7e404 ("tracing/kprobes: Return EADDRNOTAVAIL when func matches several symbols")
in the upstream kernel introduced a regression where kprobes cannot be
created on functions residing in loadable modules if the probe location
is identified by an unqualified function name.

The faulty commit was backported e.g. to kernel 6.1.60. The backported
fix is on track to be released in kernel 6.1.71. Cherry-pick the fix
until 6.1.71 is released and our upstream catches up to it.

While the faulty commit was backported to the 5.15 series as well, the
backport has not hit our upstream yet. There is nothing to be done for
our other kernel packages (but to be vigilant about not picking up a
release with the faulty backport in the near future).

Signed-off-by: Markus Boehme <markubo@amazon.com>