kernel-5.15: backport fix for creating kprobes using unqualified names
Commit b022f0c7e404 ("tracing/kprobes: Return EADDRNOTAVAIL when func matches several symbols")
in the upstream kernel introduced a regression where kprobes cannot be
created on functions residing in loadable modules if the probe location
is identified by an unqualified function name.

The faulty commit was backported to the 5.15, but a backport of the fix
is not yet available. Carry the fix here for release preparation and
seek resolution upstream as soon as possible. This may mean upstreaming,
or reverting the faulty commit entirely (sentiment for the faulty commit
has soured, and it has been yanked already from patch queues for older
stable series).

Signed-off-by: Markus Boehme <markubo@amazon.com>