kernel-6.1: merge platform configs
Remove the need for conditional compilation by merging the platform-
specific config snippets together.

Storage drivers that were set to "Y" in any platform config are now
"Y" for all platforms. Network drivers that were set to "M" in any
platform config are likewise now "M" in the merged config.

The simple DRM driver was disabled for `metal` and `vmware` but
enabled for `aws`; it is now enabled for all platforms. The i8042
driver was disabled for `aws` but enabled for `metal` and `vmware`;
it is also now enabled for all platforms.

Switch to a full list of all packaged modules to clarify the full set
of available modules. This also paves the way to move certain modules
into platform-specific subpackages in a subsequent change.

Signed-off-by: Ben Cressey <bcressey@amazon.com>