kernel-6.1: also provide uncompressed devel files
When using erofs as the root filesystem, we want to avoid other forms
of compression, such as the squashfs used for the kernel development
files.

Include the uncompressed version of these files when the erofs image
feature is set, and override the configuration for the mount unit to
do a simple bind mount instead.

Signed-off-by: Ben Cressey <bcressey@amazon.com>