kmod-5.10-nvidia: add nvidia-persistenced
Whenever the NVIDIA device resources are no longer in use, the NVIDIA
kernel driver will tear down the device state. `nvidia-persistenced`
activates persistence mode, which keeps the device files open which
prevents the kernel from removing the device state. This is desirable
in applications that may suffer performance hits due to repeated
device initialization.

The NVIDIA device drivers ship with templates for running
`nvidia-persistenced` as a systemd unit. This change uses that
template.

The `nvidia-persistenced` documentation advises that while the systemd
unit can run as root, the unit should provide a non-root user for
`nvidia-persistenced` to run under.

See the documentation included with the NVIDIA driver for more
information about `nvidia-persistenced`.