kernel: build external neuron kmod
Build the external Neuron kmod as part of the kernel build, so it can
be signed with the ephemeral module signing key. That allows it to be
loaded at runtime when kernel lockdown is in effect.

Since autoload doesn't work for this module, add a custom instance of
the modprobe unit that only runs if Neuron hardware is detected, and
run it as part of sysinit.target.

Signed-off-by: Ben Cressey <bcressey@amazon.com>