grub: add patches for Boot Loader Interface
Backport upstream patches from the series that added the Boot Loader
Interface module, which implements a different part of the interface:
support for LoaderInfo and LoaderDevicePartUUID.

The refactoring commits added a new grub_efi_set_variable_to_string
function, along with the ability to set a variable without the non-
volatile flag applied.

Add two patches of our own: one to implement the part of BLI that's
used by `systemd-analyze` to report time spent in the firmware and in
the bootloader; and another to make it so that grub_get_time_ms() on
i386 doesn't treat the time the TSC was calibrated as the epoch for
all timestamps, which otherwise hides the time spent between power on
and when GRUB is started.

Signed-off-by: Ben Cressey <bcressey@amazon.com>